feat: quizzes 페이지 코드 분리(#124)
- quizzes 페이지 코드 분리
